pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "personal-ai-employee"
version = "0.1.0"
description = "A local automation system where Claude Code acts as an AI employee that monitors an Obsidian markdown vault, processes incoming tasks, and writes structured outputs"
readme = "README.md"
authors = [{name = "Developer", email = "developer@example.com"}]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "watchdog>=4.0.0",
    "python-dotenv>=1.0.0",
    "google-genai>=0.1.0",
    "google-api-python-client>=2.70.0",
    "google-auth-oauthlib>=0.8.0",
    "linkedin-api>=0.1.0",
    "playwright>=1.30.0"
]
requires-python = ">=3.13"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0"
]

[project.scripts]
ai-employee = "personal_ai_employee.cli:main"

[tool.setuptools.packages.find]
where = ["src"]